style(story): Rename the stories file
It contains multiple stories.
