+++ src/components/Bio/Bio.test.ts
@@ -2,7 +2,7 @@ import { experimental_AstroContainer as AstroContainer } from "astro/container";
 import { describe, expect, test } from "vitest";
 
 import Bio from "./Bio.astro";
-import * as stories from "./story.ts";
+import * as stories from "./stories.ts";
 
 describe("Bio", () => {
   for (const [name, props] of Object.entries(stories)) {
