+++ src/components/Card/Card.stories.ts
@@ -1,5 +1,5 @@
 import Card from "./Card.astro";
-import * as stories from "./story.ts";
+import * as stories from "./stories.ts";
 
 export default {
   title: "Card",
