+++ src/components/SNS/SNS.test.ts
@@ -2,7 +2,7 @@ import { experimental_AstroContainer as AstroContainer } from "astro/container";
 import { describe, expect, test } from "vitest";
 
 import SNS from "./SNS.astro";
-import * as stories from "./story.ts";
+import * as stories from "./stories.ts";
 
 describe("SNS", () => {
   describe("Valid service and ID", () => {
